Fix JS syntax in post
Previous behavior assigned reference of original array to new array. This affected the demo. since adding a value-type to the "new" array also adds it to the "original" array.

New behavior creates a new array, but still makes a shallow copy of data, so the example related to objects still reinforces the idea of "pass-by-ref".